* # IMPROVEMENT #
* Adicionado confirmacao na exclusao de relatorio
* Removido softdelete de entradas no BD de relatorios e forcado exclusao
    do arquivo xlsx correspondente

+++ app/Http/Controllers/Api/AnalyticsReportController.php
@@ -103,6 +103,15 @@ class AnalyticsReportController extends BaseController
             if (!$analyticReport = AnalyticsReport::find($id)) {
                 throw new NotFoundHttpException('Recurso não encontrado');
             }
+
+            $basename = $analyticReport->filename;
+            $basepath = config('filesystems.disks.local.root');
+            $export_dir = $basepath . '/excel_exports';
+            $filename = $export_dir . '/' . $basename;
+            if (!is_file($filename)) {
+                return $this->sendError("Requisição inválida!", 404);
+            }
+            unlink($filename);
             $analyticReport->delete();
             return $this->sendResponse([], "Relatório excluído com sucesso!");
         } catch (\Exception $e) {
